Remove test data directories and generate them in test
