--- test_data/lst/users/anonymous/example-files/tmpdir/.gitignore
@@ -1,1 +1,0 @@
-*
