--- test_data/md5sum-lst.txt
@@ -1,1 +1,0 @@
-Test
